Added ChatRows

+++ Client/app/src/main/java/com/medium/client/data/remote/adapters/MoshiInstantAdapter.kt
@@ -2,6 +2,7 @@ package com.medium.client.data.remote.adapters
 
 import android.annotation.SuppressLint
 import com.squareup.moshi.*
+import java.math.BigDecimal
 import java.time.Instant
 import javax.inject.Inject
 
@@ -11,7 +12,7 @@ class MoshiInstantAdapter @Inject constructor() : JsonAdapter<Instant?>() {
     @FromJson
     override fun fromJson(reader: JsonReader): Instant? {
         return reader.readJsonValue()?.toString()?.let {
-            val milliseconds = it.toLong()
+            val milliseconds = it.toBigDecimal().toLong()
             Instant.ofEpochMilli(milliseconds)
         }
     }
